requirements.txt
click==6.7
Flask==0.12.2
Flask-Script==2.0.5
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.13
Flask-SQLAlchemy==2.2
psycopg2==2.7.1
Flask-Testing==0.6.2
gunicorn==19.7.1
coverage==4.4.1
flask-cors==3.0.2
flask-migrate==2.0.4
flask-bcrypt==0.7.1
pyjwt==1.5.0